# Conflicts:
#	src/main/java/com/netcracker/service/request/RequestService.java
#	src/main/java/com/netcracker/service/request/RequestServiceImpl.java

+++ src/main/java/com/netcracker/service/request/RequestServiceImpl.java
@@ -1,17 +1,8 @@
 package com.netcracker.service.request;
 
 import com.netcracker.exception.*;
-import com.netcracker.model.entity.Person;
-import com.netcracker.model.entity.Priority;
-import com.netcracker.model.entity.Request;
-import com.netcracker.model.entity.Status;
-import com.netcracker.repository.common.Pageable;
-import com.netcracker.repository.data.impl.RequestRepositoryImpl;
-import com.netcracker.repository.data.interfaces.PersonRepository;
-import com.netcracker.repository.data.interfaces.PriorityRepository;
-import com.netcracker.repository.data.interfaces.RequestRepository;
-import com.netcracker.repository.data.interfaces.StatusRepository;
-import com.netcracker.service.person.PersonService;
+import com.netcracker.model.entity.*;
+import com.netcracker.repository.data.interfaces.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
@@ -19,27 +10,42 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.security.Principal;
 import java.sql.Timestamp;
-import java.util.Date;
-import java.util.List;
-import java.util.Optional;
+import java.util.*;
 
 @Service
 public class RequestServiceImpl implements RequestService {
 
-    @Autowired
-    private RequestRepository requestRepository;
+    private final RequestRepository requestRepository;
+
+    private final PersonRepository personRepository;
+
+    private final StatusRepository statusRepository;
+
+    private final RoleRepository roleRepository;
+
+    private final PriorityRepository priorityRepository;
+
+    private final ChangeGroupRepository changeGroupRepository;
+
+    private final FieldRepository fieldRepository;
+
 
     @Autowired
-    private PersonService personService;
-
-    @Autowired
-    private StatusRepository statusRepository;
-
-    @Autowired
-    private PriorityRepository priorityRepository;
-
-    @Autowired
-    private PersonRepository personRepository;
+    public RequestServiceImpl(RequestRepository requestRepository,
+                              PersonRepository personRepository,
+                              StatusRepository statusRepository,
+                              RoleRepository roleRepository,
+                              PriorityRepository priorityRepository,
+                              ChangeGroupRepository changeGroupRepository,
+                              FieldRepository fieldRepository) {
+        this.requestRepository = requestRepository;
+        this.personRepository = personRepository;
+        this.statusRepository = statusRepository;
+        this.roleRepository = roleRepository;
+        this.priorityRepository = priorityRepository;
+        this.changeGroupRepository = changeGroupRepository;
+        this.fieldRepository = fieldRepository;
+    }
 
     @Override
     public Optional<Request> getRequestById(Long id) {
@@ -50,33 +56,43 @@ public class RequestServiceImpl implements RequestService {
 
     @Transactional(propagation = Propagation.REQUIRED)
     @Override
-    public Optional<Request> saveSubRequest(Request subRequest) throws CannotCreateSubRequestException {
-        if (subRequest.getParent() != null) {
-            Request parentRequest = requestRepository.findOne(subRequest.getParent().getId()).orElseThrow(() ->
-                    new CannotCreateSubRequestException("No parent request with id " + subRequest.getParent().getId()));
-
-            if (parentRequest.getParent() != null) {
-                throw new CannotCreateSubRequestException("Parent request is sub request");
-            }
-
-            String parentStatus = parentRequest.getStatus().getName();
-            if ("CANCELED".equals(parentStatus) || "CLOSED".equals(parentStatus)) {
-                throw new CannotCreateSubRequestException("Parent request is closed or canceled");
-            }
-
-            subRequest.setCreationTime(new Timestamp(System.currentTimeMillis()));
-            subRequest.setEmployee(parentRequest.getEmployee());
-            subRequest.setPriority(parentRequest.getPriority());
-            subRequest.setStatus(statusRepository.findStatusByName("IN PROGRESS").orElseThrow(() ->
-                    new CannotCreateSubRequestException("No status 'IN PROGRESS'")));
-            return requestRepository.save(subRequest);
-        } else {
+    public Optional<Request> saveSubRequest(Request subRequest, String email) throws CannotCreateSubRequestException {
+        if (subRequest.getParent() == null) {
             throw new CannotCreateSubRequestException("No parent request");
         }
+
+        Person manager = personRepository.findPersonByEmail(email).orElseThrow(() ->
+                new CannotCreateSubRequestException("No manager with email " + email));
+
+        subRequest.setManager(manager);
+
+        Request parentRequest = requestRepository.findOne(subRequest.getParent().getId()).orElseThrow(() ->
+                new CannotCreateSubRequestException("No parent request with id " + subRequest.getParent().getId()));
+
+        if (parentRequest.getParent() != null) {
+            throw new CannotCreateSubRequestException("Parent request is sub request");
+        }
+
+        String parentStatus = parentRequest.getStatus().getName();
+        if ("CANCELED".equals(parentStatus) || "CLOSED".equals(parentStatus)) {
+            throw new CannotCreateSubRequestException("Parent request is closed or canceled");
+        }
+
+        subRequest.setCreationTime(new Timestamp(System.currentTimeMillis()));
+        subRequest.setEmployee(parentRequest.getEmployee());
+        subRequest.setPriority(parentRequest.getPriority());
+        subRequest.setStatus(statusRepository.findStatusByName("IN PROGRESS").orElseThrow(() ->
+                new CannotCreateSubRequestException("No status 'IN PROGRESS'")));
+        return requestRepository.save(subRequest);
     }
 
     @Override
-    public Optional<Request> saveRequest(Request request) throws CannotCreateRequestException {
+    public Optional<Request> saveRequest(Request request, String email) throws CannotCreateRequestException {
+        Person manager = personRepository.findPersonByEmail(email).orElseThrow(() ->
+                new CannotCreateRequestException("No employee with email " + email));
+
+        request.setEmployee(manager);
+
         request.setStatus(statusRepository.findStatusByName("FREE").orElseThrow(() ->
                 new CannotCreateRequestException("No status 'FREE'")
         ));
@@ -85,7 +101,8 @@ public class RequestServiceImpl implements RequestService {
     }
 
     @Override
-    public Optional<Request> updateRequest(Request request) {
+    public Optional<Request> updateRequest(Request request, Long requestId) {
+        request.setId(requestId);
         return this.requestRepository.updateRequest(request);
     }
 
@@ -102,7 +119,11 @@ public class RequestServiceImpl implements RequestService {
     @Transactional
     @Override
     public void deleteRequestById(Long id) throws CannotDeleteRequestException, ResourceNotFoundException {
-        Request request = getRequestById(id).get();
+        Optional<Request> requestOptional = getRequestById(id);
+        if(!requestOptional.isPresent()) {
+            throw new CannotDeleteRequestException("No such request id " + id);
+        }
+        Request request = requestOptional.get();
         if (request.getParent() != null) {
             this.requestRepository.delete(id);
         } else {
@@ -123,6 +144,18 @@ public class RequestServiceImpl implements RequestService {
     @Override
     public int changeRequestStatus(Request request, Status status) {
         return requestRepository.changeRequestStatus(request, status);
+    }
+
+    @Override
+    @Transactional(readOnly = true)
+    public Set<ChangeGroup> getRequestHistory(Long requestId, String period) {
+        try {
+            Set<ChangeGroup> changeGroups = changeGroupRepository.findByRequestIdWithDetails(requestId,Period.valueOf(period.toUpperCase()));
+            fill(changeGroups);
+            return changeGroups;
+        }catch (IllegalArgumentException e){
+            return new HashSet<ChangeGroup>();
+        }
     }
 
     @Override
@@ -169,13 +202,17 @@ public class RequestServiceImpl implements RequestService {
     private void fill(Request request) {
         Person employee = request.getEmployee();
         if(employee != null) {
-            employee = personService.getPersonById(employee.getId()).orElseGet(null);
+            employee = personRepository.findOne(employee.getId()).orElseGet(null);
+
+            Role role = roleRepository.findOne(employee.getRole().getId()).orElseGet(null);
+            employee.setRole(role);
+
             request.setEmployee(employee);
         }
 
         Person manager = request.getManager();
         if(manager != null) {
-            manager = personService.getPersonById(manager.getId()).orElseGet(null);
+            manager = personRepository.findOne(manager.getId()).orElseGet(null);
             request.setManager(manager);
         }
 
@@ -184,5 +221,11 @@ public class RequestServiceImpl implements RequestService {
 
         Status status = request.getStatus();
         request.setStatus(statusRepository.findOne(status.getId()).orElseGet(null));
+    }
+
+    private void fill(Set<ChangeGroup>  changeGroup){
+        changeGroup.forEach(cg->cg.getChangeItems().forEach(ci->{
+                    ci.setField(fieldRepository.findOne(ci.getField().getId()).get());
+                }));
     }
 }
